requirements.txt
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.21.0
requests>=2.25.0
unstructured[all]
regex>=2023.10.3
tree-sitter==0.20.4
tree-sitter-languages==1.8.0
jira
gitpython
python-docx
lxml
streamlit
pyinstaller
atlassian-python-api
requests
beautifulsoup4
openai
httpx
PyYAML
